Split MapService file
