+++ src/components/feedback/FeedbackDirective.js
@@ -1,13 +1,13 @@
 goog.provide('ga_feedback_directive');
 
 goog.require('ga_browsersniffer_service');
-goog.require('ga_export_kml_service');
+goog.require('ga_exportkml_service');
 goog.require('ga_permalink');
 (function() {
 
   var module = angular.module('ga_feedback_directive', [
     'ga_browsersniffer_service',
-    'ga_export_kml_service',
+    'ga_exportkml_service',
     'ga_permalink',
     'pascalprecht.translate'
   ]);
