+++ src/components/layermanager/LayermanagerDirective.js
@@ -1,7 +1,7 @@
 goog.provide('ga_layermanager_directive');
 
 goog.require('ga_attribution_service');
-goog.require('ga_layer_metadata_popup_service');
+goog.require('ga_layermetadatapopup_service');
 goog.require('ga_map_service');
 goog.require('ga_urlutils_service');
 
@@ -9,7 +9,7 @@ goog.require('ga_urlutils_service');
 
   var module = angular.module('ga_layermanager_directive', [
     'pascalprecht.translate',
-    'ga_layer_metadata_popup_service',
+    'ga_layermetadatapopup_service',
     'ga_map_service',
     'ga_attribution_service',
     'ga_urlutils_service'
