+++ src/components/permalink/PermalinkService.js
@@ -131,5 +131,4 @@ goog.require('ga_urlutils_service');
         return permalink;
       };
   });
-
 })();
