+++ src/js/GaModule.js
@@ -47,7 +47,7 @@ goog.require('ga_seo');
 goog.require('ga_seo_controller');
 goog.require('ga_share');
 goog.require('ga_share_controller');
-goog.require('ga_styles_from_literals_service');
+goog.require('ga_stylesfromliterals_service');
 goog.require('ga_swipe');
 goog.require('ga_tilt3d');
 goog.require('ga_timeselector');
@@ -91,7 +91,7 @@ goog.require('ga_waitcursor_service');
     'ga_profile',
     'ga_fullscreen',
     'ga_waitcursor_service',
-    'ga_styles_from_literals_service',
+    'ga_stylesfromliterals_service',
     'ga_seo',
     'ga_draw',
     'ga_query',
